refact: payment 페이지 payment/[userId]로 수정 (#99)
