+++ src/app/cart/[userId]/page.tsx
@@ -3,10 +3,8 @@ import React, { useState } from "react";
 import Section from "@/components/layout/Section";
 import CartItem from "../CartItem";
 import { useCart } from "@/hooks";
-import Link from "next/link";
 import PageBreadCrumb from "@/components/layout/PageBreadCrumb";
 import { CartBox } from "@/types/db";
-import useLogedInStore from "@/store/logedStore";
 import { useRouter } from "next/navigation";
 
 const linkList = [
@@ -22,8 +20,6 @@ const linkList = [
 
 const CartPage = ({ params }: { params: { userId: string } }) => {
   const router = useRouter();
-  //로그인 여부 확인
-  const { logedIn } = useLogedInStore();
 
   //userId
   const userId = params.userId;
@@ -38,64 +34,55 @@ const CartPage = ({ params }: { params: { userId: string } }) => {
     <>
       <PageBreadCrumb linkList={linkList} />
       <Section title={"장바구니"} isCenter={true}>
-        {logedIn ? (
-          !isLoading ? (
-            cart !== undefined ? (
-              cart.length !== 0 ? (
-                <div>
-                  <div className="flex flex-col items-center justify-center">
-                    {(cart as CartBox[]).map((cartItem, idx) => {
-                      return (
-                        cartItem && (
-                          <CartItem
-                            cart={cartItem}
-                            cartPrice={cartPrice}
-                            setCartPrice={setCartPrice}
-                            originPrice={originPrice}
-                            setOriginPrice={setOriginPrice}
-                            idx={idx}
-                            key={cartItem.id}
-                          />
-                        )
-                      );
-                    })}
-                  </div>
-
-                  <div className="mt-[80px] flex flex-col content-end">
-                    <div className="mb-[30px]">
-                      <div className="text-right text-[16px] font-bold mr-1">
-                        총 결제금액
-                        <p className="text-[24px] font-bold inline-block ml-1">
-                          {" "}
-                          {cartPrice.reduce((acc, num) => acc + num, 0)}원
-                        </p>
-                      </div>
-                    </div>
-
-                    <button
-                      onClick={() => router.push("/payment")}
-                      className="w-[244px] h-[50px] bg-point text-white rounded-[5px] ml-[80%]"
-                    >
-                      결제하기
-                    </button>
-                  </div>
+        {!isLoading ? (
+          cart !== undefined ? (
+            cart.length !== 0 ? (
+              <div>
+                <div className="flex flex-col items-center justify-center">
+                  {(cart as CartBox[]).map((cartItem, idx) => {
+                    return (
+                      cartItem && (
+                        <CartItem
+                          cart={cartItem}
+                          cartPrice={cartPrice}
+                          setCartPrice={setCartPrice}
+                          originPrice={originPrice}
+                          setOriginPrice={setOriginPrice}
+                          idx={idx}
+                          key={cartItem.id}
+                        />
+                      )
+                    );
+                  })}
                 </div>
-              ) : (
-                <div>카트가 비어있습니다</div>
-              )
+
+                <div className="mt-[80px] flex flex-col content-end">
+                  <div className="mb-[30px]">
+                    <div className="text-right text-[16px] font-bold mr-1">
+                      총 결제금액
+                      <p className="text-[24px] font-bold inline-block ml-1">
+                        {" "}
+                        {cartPrice.reduce((acc, num) => acc + num, 0)}원
+                      </p>
+                    </div>
+                  </div>
+
+                  <button
+                    onClick={() => router.push(`/payment/${userId}`)}
+                    className="w-[244px] h-[50px] bg-point text-white rounded-[5px] ml-[80%]"
+                  >
+                    결제하기
+                  </button>
+                </div>
+              </div>
             ) : (
-              <div>카트 불러오는 중...</div>
+              <div>카트가 비어있습니다</div>
             )
           ) : (
-            <div>Loading...</div>
+            <div>카트 불러오는 중...</div>
           )
         ) : (
-          <div className="flex flex-col items-center justify-center space-y-5">
-            <p>로그인 후 이용하세요</p>
-            <Link href={"/auth"} className="text-[18px] text-point font-bold">
-              로그인 하기
-            </Link>
-          </div>
+          <div>Loading...</div>
         )}
       </Section>
     </>
